backend/requirements.txt
Flask
python-dotenv
openai
python-docx
PyPDF2
